+++ src/pkg-core/IdElements/TypeId.ts
@@ -86,6 +86,7 @@ namespace TypeId {
 	export const EmbedOnly = [
 		'ability',
 		'condition',
+		'outcome',
 		'option',
 		'row',
 		'feature',
@@ -99,6 +100,7 @@ namespace TypeId {
 		'move',
 		'ability',
 		'condition',
+		'outcome',
 		'option',
 		'row',
 		'feature',
@@ -113,7 +115,7 @@ namespace TypeId {
 		ability: ['move', 'oracle_rollable'],
 		truth: ['option'],
 		option: ['oracle_rollable'],
-		move: ['condition', 'oracle_rollable'],
+		move: ['condition', 'outcome', 'oracle_rollable'],
 		oracle_rollable: ['row'],
 		delve_site: ['denizen'],
 		delve_site_domain: ['feature', 'danger'],
@@ -128,7 +130,7 @@ namespace TypeId {
 	/** Types that can be an embed of an embed. */
 	export const EmbeddableInEmbeddedTypeMap = {
 		ability: ['oracle_rollable', 'move'],
-		move: ['condition'],
+		move: ['condition', 'outcome'],
 		option: ['oracle_rollable'],
 		oracle_rollable: ['row'],
 	} as const satisfies {
@@ -205,6 +207,7 @@ namespace TypeId {
 	export const EmbeddedPropertyKeys = {
 		ability: 'abilities',
 		condition: 'trigger.conditions',
+		outcome: 'outcomes',
 		option: 'options',
 		row: 'rows',
 		feature: 'features',
@@ -219,6 +222,7 @@ namespace TypeId {
 	const EmbeddedPropertyType = {
 		abilities: 'array',
 		'trigger.conditions': 'array',
+		outcomes: 'dictionary',
 		dangers: 'array',
 		denizens: 'array',
 		features: 'array',
@@ -264,6 +268,7 @@ namespace TypeId {
 		'truth',
 		'ability',
 		'condition',
+		'outcome',
 		'option',
 		'row',
 		'feature',